rust: use bindgen --size_t-is-usize option
This will give us identical bindings for all platforms regardless of
what underlying C type they use to define size_t.

For some reason, this isn't the default:
https://github.com/rust-lang/rust-bindgen/issues/1901

Previously, we generated different things on different platforms:

x86-64:     pub type size_t = ::std::os::raw::c_ulong;
32-bit arm: pub type size_t = ::std::os::raw::c_uint;

It also lets us eliminate a couple of size_t casts in the Rust wrapper
code.

Bug: None
Test: (cd rust/minijail; cargo test)

+++ rust/minijail/src/lib.rs
@@ -465,7 +465,7 @@ impl Minijail {
     }
     pub fn set_supplementary_gids(&mut self, ids: &[libc::gid_t]) {
         unsafe {
-            minijail_set_supplementary_gids(self.jail, ids.len() as size_t, ids.as_ptr());
+            minijail_set_supplementary_gids(self.jail, ids.len(), ids.as_ptr());
         }
     }
     pub fn keep_supplementary_gids(&mut self) {
@@ -767,7 +767,7 @@ impl Minijail {
     }
     pub fn mount_tmp_size(&mut self, size: usize) {
         unsafe {
-            minijail_mount_tmp_size(self.jail, size as size_t);
+            minijail_mount_tmp_size(self.jail, size);
         }
     }
     pub fn mount_bind<P1: AsRef<Path>, P2: AsRef<Path>>(
